ReCollect/src/main/java/ua/tqs/ReCollect/repository/ItemRepository.java
package ua.tqs.ReCollect.repository;

import java.util.List;
import java.util.Optional;

import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

import ua.tqs.ReCollect.model.Item;

@Repository
public interface ItemRepository extends JpaRepository<Item, Long> {

    public List<Item> findAll();

    public Optional<Item> findById(Long id);

    public List<Item> findByName(String name);

    public void deleteById(Long id);

    public List<Item> findTop20ByOrderByCreationDateAsc();

    public List<Item> findTop20ByOrderByCreationDateDesc();

}



    

